+++ src/com/dandymadeproductions/ajqvue/gui/panels/TableTabPanel_PostgreSQL.java
@@ -13,7 +13,7 @@
 //
 //==============================================================
 // Copyright (C) 2016-2018 Dana M. Proctor
-// Version 1.4 06/21/2018
+// Version 1.5 07/04/2018
 //
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -49,6 +49,8 @@
 //             System.out to type name.
 //         1.4 Methods addItem() & editSelectedItem() Change in Arguments for
 //             TableEntryForm to Meet New Constructor Requirments.
+//         1.5 Methods getColumnNames() & loadTable(), columnSQLTypeHashMap
+//             Proper Loading of Integer & Extracting int.
 //             
 //-----------------------------------------------------------------
 //                  danap@dandymadeproductions.com
@@ -81,7 +83,7 @@ import com.dandymadeproductions.ajqvue.utilities.Utils;
  * mechanism to page through the database table's data.
  * 
  * @author Dana M. Proctor
- * @version 1.3 06/21/2018
+ * @version 1.5 07/04/2018
  */
 
 public class TableTabPanel_PostgreSQL extends TableTabPanel //implements ActionListener
@@ -196,7 +198,7 @@ public class TableTabPanel_PostgreSQL extends TableTabPanel //implements ActionL
             colNameString = tableMetaData.getColumnName(i);
             comboBoxNameString = parseColumnNameField(colNameString);
             columnClass = tableMetaData.getColumnClassName(i);
-            columnSQLType = tableMetaData.getColumnType(i);
+            columnSQLType = Integer.valueOf(tableMetaData.getColumnType(i));
             columnTypeName = tableMetaData.getColumnTypeName(i);
             columnSize = Integer.valueOf(tableMetaData.getColumnDisplaySize(i));
 
@@ -536,7 +538,7 @@ public class TableTabPanel_PostgreSQL extends TableTabPanel //implements ActionL
                String currentHeading = headings.next();
                columnName = columnNamesHashMap.get(currentHeading);
                columnClass = columnClassHashMap.get(currentHeading);
-               columnSQLType = columnSQLTypeHashMap.get(currentHeading);
+               columnSQLType = columnSQLTypeHashMap.get(currentHeading).intValue();
                columnTypeName = columnTypeNameHashMap.get(currentHeading);
                columnSize = (columnSizeHashMap.get(currentHeading)).intValue();
                keyLength = keyLengthHashMap.get(columnName);
